// src/app/shared/interfaces/placeholderAllocation.interface.ts
export interface PlaceholderAllocation {
  id?: string;
  planningCardId?: string;
  planningCardTitle?: string;
  oldCaseCode: string;
  caseName: string;
  caseTypeCode?: number;
  clientName: string;
  pipelineId: string;
  opportunityName: string;
  employeeCode: string;
  employeeName: string;
  internetAddress?: string;
  operatingOfficeCode: number;
  operatingOfficeAbbreviation: string;
  currentLevelGrade: string;
  serviceLineCode: string;
  serviceLineName: string;
  allocation: number;
  joiningDate?: string;
  startDate: string;
  endDate: string;
  investmentCode: number;
  investmentName: string;
  caseRoleCode: string;
  lastUpdatedBy: string;
  caseStartDate?: string;
  caseEndDate?: string;
  opportunityStartDate?: string;
  opportunityEndDate?: string;
  notes?: string;
  isPlaceholderAllocation?: boolean;
  commitmentTypeCode?: string;
  commitmentTypeName?: string;
  isConfirmed?: boolean;
  probabilityPercent?: number;
  positionGroupCode?: string;
}